Crate anvil

Modules

cmd

contains cli command

config 🔒

eth

ethereum related implementations

filter

support for polling filters Support for polling based filters

hardfork 🔒

logging

commandline output User facing Logger

pubsub

types for subscriptions

server

axum RPC server implementations Contains the code to launch an ethereum RPC-Server

service 🔒

contains the background service that drives the node background service

shutdown 🔒

Futures for shutdown signal Helper for shutdown signals

tasks 🔒

additional task management Task management support

Structs

AccountGenerator

Can create dev accounts

NodeConfig

Configurations of the EVM node

NodeHandle

A handle to the spawned node and server tasks

Enums

Hardfork

Constants

CHAIN_ID

Default chain id of the node

VERSION_MESSAGE

anvil 0.1.0 (f01b232bc 2022-04-13T23:28:39.493201+00:00)

Functions

spawn

Creates the node and runs the server

Type Aliases

IpcTask 🔒