anvil
Module cmd

Available on crate feature cmd only.

contains cli command

Structs

AnvilEvmArgs

Anvil’s EVM related arguments.

ForkUrl

Represents the input URL for a fork with an optional trailing block number: http://localhost:8545@1000000

NodeArgs

PeriodicStateDumper 🔒

Helper type to periodically dump the state of the chain to disk

StateFile

Represents the –state flag and where to load from, or dump the state to

Constants

DEFAULT_DUMP_INTERVAL 🔒

Default interval for periodically dumping the state.

IPC_HELP 🔒 Non-Windows

The default IPC endpoint

Functions

duration_from_secs_f64 🔒

read_genesis_file 🔒

Clap’s value parser for genesis. Loads a genesis.json file.