anvil::config
Struct NodeConfig

pub struct NodeConfig {

    pub chain_id: Option<u64>,
    pub gas_limit: Option<u128>,
    pub disable_block_gas_limit: bool,
    pub gas_price: Option<u128>,
    pub base_fee: Option<u64>,
    pub disable_min_priority_fee: bool,
    pub blob_excess_gas_and_price: Option<BlobExcessGasAndPrice>,
    pub hardfork: Option<ChainHardfork>,
    pub genesis_accounts: Vec<PrivateKeySigner>,
    pub genesis_balance: U256,
    pub genesis_timestamp: Option<u64>,
    pub signer_accounts: Vec<PrivateKeySigner>,
    pub block_time: Option<Duration>,
    pub no_mining: bool,
    pub mixed_mining: bool,
    pub port: u16,
    pub max_transactions: usize,
    pub eth_rpc_url: Option<String>,
    pub fork_choice: Option<ForkChoice>,
    pub fork_headers: Vec<String>,
    pub fork_chain_id: Option<U256>,
    pub account_generator: Option<AccountGenerator>,
    pub enable_tracing: bool,
    pub no_storage_caching: bool,
    pub server_config: ServerConfig,
    pub host: Vec<IpAddr>,
    pub transaction_order: TransactionOrder,
    pub config_out: Option<String>,
    pub genesis: Option<Genesis>,
    pub fork_request_timeout: Duration,
    pub fork_request_retries: u32,
    pub fork_retry_backoff: Duration,
    pub compute_units_per_second: u64,
    pub ipc_path: Option<Option<String>>,
    pub enable_steps_tracing: bool,
    pub print_logs: bool,
    pub enable_auto_impersonate: bool,
    pub code_size_limit: Option<usize>,
    pub prune_history: PruneStateHistoryConfig,
    pub max_persisted_states: Option<usize>,
    pub init_state: Option<SerializableState>,
    pub transaction_block_keeper: Option<usize>,
    pub disable_default_create2_deployer: bool,
    pub enable_optimism: bool,
    pub slots_in_an_epoch: u64,
    pub memory_limit: Option<u64>,
    pub precompile_factory: Option<Arc<dyn PrecompileFactory>>,
    pub alphanet: bool,
    pub silent: bool,

}

Configurations of the EVM node

Fields
chain_id: Option<u64>

Chain ID of the EVM chain

gas_limit: Option<u128>

Default gas limit for all txs

disable_block_gas_limit: bool

If set to true, disables the block gas limit

gas_price: Option<u128>

Default gas price for all txs

base_fee: Option<u64>

Default base fee

disable_min_priority_fee: bool

If set to true, disables the enforcement of a minimum suggested priority fee

blob_excess_gas_and_price: Option<BlobExcessGasAndPrice>

Default blob excess gas and price

hardfork: Option<ChainHardfork>

The hardfork to use

genesis_accounts: Vec<PrivateKeySigner>

Signer accounts that will be initialised with genesis_balance in the genesis block

genesis_balance: U256

Native token balance of every genesis account in the genesis block

genesis_timestamp: Option<u64>

Genesis block timestamp

signer_accounts: Vec<PrivateKeySigner>

Signer accounts that can sign messages/transactions from the EVM node

block_time: Option<Duration>

Configured block time for the EVM chain. Use None to mine a new block for every tx

no_mining: bool

Disable auto, interval mining mode uns use MiningMode::None instead

mixed_mining: bool

Enables auto and interval mining mode

port: u16

port to use for the server

max_transactions: usize

maximum number of transactions in a block

eth_rpc_url: Option<String>

url of the rpc server that should be used for any rpc calls

fork_choice: Option<ForkChoice>

pins the block number or transaction hash for the state fork

fork_headers: Vec<String>

headers to use with eth_rpc_url

fork_chain_id: Option<U256>

specifies chain id for cache to skip fetching from remote in offline-start mode

account_generator: Option<AccountGenerator>

The generator used to generate the dev accounts

enable_tracing: bool

whether to enable tracing

no_storage_caching: bool

Explicitly disables the use of RPC caching.

server_config: ServerConfig

How to configure the server

host: Vec<IpAddr>

The host the server will listen on

transaction_order: TransactionOrder

How transactions are sorted in the mempool

config_out: Option<String>

Filename to write anvil output as json

genesis: Option<Genesis>

The genesis to use to initialize the node

fork_request_timeout: Duration

Timeout in for requests sent to remote JSON-RPC server in forking mode

fork_request_retries: u32

Number of request retries for spurious networks

fork_retry_backoff: Duration

The initial retry backoff

compute_units_per_second: u64

available CUPS

ipc_path: Option<Option<String>>

The ipc path

enable_steps_tracing: bool

Enable transaction/call steps tracing for debug calls returning geth-style traces

print_logs: bool

Enable printing of console.log invocations.

enable_auto_impersonate: bool

Enable auto impersonation of accounts on startup

code_size_limit: Option<usize>

Configure the code size limit

prune_history: PruneStateHistoryConfig

Configures how to remove historic state.

If set to Some(num) keep latest num state in memory only.

max_persisted_states: Option<usize>

Max number of states cached on disk.

init_state: Option<SerializableState>

The file where to load the state from

transaction_block_keeper: Option<usize>

max number of blocks with transactions in memory

disable_default_create2_deployer: bool

Disable the default CREATE2 deployer

enable_optimism: bool

Enable Optimism deposit transaction

slots_in_an_epoch: u64

Slots in an epoch

memory_limit: Option<u64>

The memory limit per EVM execution in bytes.

precompile_factory: Option<Arc<dyn PrecompileFactory>>

Factory used by anvil to extend the EVM’s precompiles.

alphanet: bool

Enable Alphanet features.

silent: bool

Do not print log messages.

Implementations

impl NodeConfig

fn as_string(&self, fork: Option<&ClientFork>) -> String

fn as_json(&self, fork: Option<&ClientFork>) -> Value

impl NodeConfig

pub fn empty_state() -> Self

Returns a new config which does not initialize any accounts on node startup.

impl NodeConfig

pub fn with_memory_limit(self, mems_value: Option<u64>) -> Self

Returns the memory limit of the node

pub fn get_base_fee(&self) -> u64

Returns the base fee to use

pub fn get_gas_price(&self) -> u128

Returns the base fee to use

pub fn get_blob_excess_gas_and_price(&self) -> BlobExcessGasAndPrice

pub fn get_hardfork(&self) -> ChainHardfork

Returns the hardfork to use

pub fn with_code_size_limit(self, code_size_limit: Option<usize>) -> Self

Sets a custom code size limit

pub fn disable_code_size_limit(self, disable_code_size_limit: bool) -> Self

Disables code size limit

pub fn with_init_state(self, init_state: Option<SerializableState>) -> Self

Sets the init state if any

pub fn with_init_state_path(self, path: impl AsRef<Path>) -> Self

Loads the init state from a file if it exists

pub fn with_chain_id<U: Into<u64>>(self, chain_id: Option<U>) -> Self

Sets the chain ID

pub fn get_chain_id(&self) -> u64

Returns the chain ID to use

pub fn set_chain_id(&mut self, chain_id: Option<impl Into<u64>>)

Sets the chain id and updates all wallets

pub fn with_gas_limit(self, gas_limit: Option<u128>) -> Self

Sets the gas limit

pub fn disable_block_gas_limit(self, disable_block_gas_limit: bool) -> Self

Disable block gas limit check

If set to true block gas limit will not be enforced

pub fn with_gas_price(self, gas_price: Option<u128>) -> Self

Sets the gas price

pub fn set_pruned_history(self, prune_history: Option<Option<usize>>) -> Self

Sets prune history status.

pub fn with_max_persisted_states<U: Into<usize>>( self, max_persisted_states: Option<U>, ) -> Self

Sets max number of states to cache on disk.

pub fn with_transaction_block_keeper<U: Into<usize>>( self, transaction_block_keeper: Option<U>, ) -> Self

Sets max number of blocks with transactions to keep in memory

pub fn with_base_fee(self, base_fee: Option<u64>) -> Self

Sets the base fee

pub fn disable_min_priority_fee(self, disable_min_priority_fee: bool) -> Self

Disable the enforcement of a minimum suggested priority fee

pub fn with_genesis(self, genesis: Option<Genesis>) -> Self

Sets the init genesis (genesis.json)

pub fn get_genesis_timestamp(&self) -> u64

Returns the genesis timestamp to use

pub fn with_genesis_timestamp<U: Into<u64>>(self, timestamp: Option<U>) -> Self

Sets the genesis timestamp

pub fn with_hardfork(self, hardfork: Option<ChainHardfork>) -> Self

Sets the hardfork

pub fn with_genesis_accounts(self, accounts: Vec<PrivateKeySigner>) -> Self

Sets the genesis accounts

pub fn with_signer_accounts(self, accounts: Vec<PrivateKeySigner>) -> Self

Sets the signer accounts

pub fn with_account_generator(self, generator: AccountGenerator) -> Self

Sets both the genesis accounts and the signer accounts so that genesis_accounts == accounts

pub fn with_genesis_balance<U: Into<U256>>(self, balance: U) -> Self

Sets the balance of the genesis accounts in the genesis block

pub fn with_blocktime<D: Into<Duration>>(self, block_time: Option<D>) -> Self

Sets the block time to automine blocks

pub fn with_mixed_mining<D: Into<Duration>>( self, mixed_mining: bool, block_time: Option<D>, ) -> Self

pub fn with_no_mining(self, no_mining: bool) -> Self

If set to true auto mining will be disabled

pub fn with_slots_in_an_epoch(self, slots_in_an_epoch: u64) -> Self

Sets the slots in an epoch

pub fn with_port(self, port: u16) -> Self

Sets the port to use

pub fn with_ipc(self, ipc_path: Option<Option<String>>) -> Self

Sets the ipc path to use

Note: this is a double Option for - None -> no ipc - Some(None) -> use default path - Some(Some(path)) -> use custom path

pub fn set_config_out(self, config_out: Option<String>) -> Self

Sets the file path to write the Anvil node’s config info to.

pub fn no_storage_caching(self) -> Self

Disables storage caching

pub fn with_storage_caching(self, storage_caching: bool) -> Self

pub fn with_eth_rpc_url<U: Into<String>>(self, eth_rpc_url: Option<U>) -> Self

Sets the eth_rpc_url to use when forking

pub fn with_fork_block_number<U: Into<u64>>( self, fork_block_number: Option<U>, ) -> Self

Sets the fork_choice to use to fork off from based on a block number

pub fn with_fork_transaction_hash<U: Into<TxHash>>( self, fork_transaction_hash: Option<U>, ) -> Self

Sets the fork_choice to use to fork off from based on a transaction hash

pub fn with_fork_choice<U: Into<ForkChoice>>( self, fork_choice: Option<U>, ) -> Self

Sets the fork_choice to use to fork off from

pub fn with_fork_chain_id(self, fork_chain_id: Option<U256>) -> Self

Sets the fork_chain_id to use to fork off local cache from

pub fn with_fork_headers(self, headers: Vec<String>) -> Self

Sets the fork_headers to use with eth_rpc_url

pub fn fork_request_timeout( self, fork_request_timeout: Option<Duration>, ) -> Self

Sets the fork_request_timeout to use for requests

pub fn fork_request_retries(self, fork_request_retries: Option<u32>) -> Self

Sets the fork_request_retries to use for spurious networks

pub fn fork_retry_backoff(self, fork_retry_backoff: Option<Duration>) -> Self

Sets the initial fork_retry_backoff for rate limits

pub fn fork_compute_units_per_second( self, compute_units_per_second: Option<u64>, ) -> Self

Sets the number of assumed available compute units per second

See also, https://docs.alchemy.com/reference/compute-units#what-are-cups-compute-units-per-second

pub fn with_tracing(self, enable_tracing: bool) -> Self

Sets whether to enable tracing

pub fn with_steps_tracing(self, enable_steps_tracing: bool) -> Self

Sets whether to enable steps tracing

pub fn with_print_logs(self, print_logs: bool) -> Self

Sets whether to print console.log invocations to stdout.

pub fn with_auto_impersonate(self, enable_auto_impersonate: bool) -> Self

Sets whether to enable autoImpersonate

pub fn with_server_config(self, config: ServerConfig) -> Self

pub fn with_host(self, host: Vec<IpAddr>) -> Self

Sets the host the server will listen on

pub fn with_transaction_order(self, transaction_order: TransactionOrder) -> Self

pub fn get_ipc_path(&self) -> Option<String>

Returns the ipc path for the ipc endpoint if any

pub fn print(&self, fork: Option<&ClientFork>)

Prints the config info

pub fn block_cache_path(&self, block: u64) -> Option<PathBuf>

Returns the path where the cache file should be stored

See also [ Config::foundry_block_cache_file()]

pub fn with_optimism(self, enable_optimism: bool) -> Self

Sets whether to enable optimism support

pub fn with_disable_default_create2_deployer(self, yes: bool) -> Self

Sets whether to disable the default create2 deployer

pub fn with_precompile_factory( self, factory: impl PrecompileFactory + 'static, ) -> Self

Injects precompiles to anvil’s EVM.

pub fn with_alphanet(self, alphanet: bool) -> Self

Sets whether to enable Alphanet support

pub fn silent(self) -> Self

Makes the node silent to not emit anything on stdout

pub fn set_silent(self, silent: bool) -> Self

pub(crate) async fn setup(&mut self) -> Backend

Configures everything related to env, backend and database and returns the Backend

Note: only memory based backend for now

pub async fn setup_fork_db( &mut self, eth_rpc_url: String, env: &mut EnvWithHandlerCfg, fees: &FeeManager, ) -> (Arc<RwLock<Box<dyn Db>>>, Option<ClientFork>)

Configures everything related to forking based on the passed eth_rpc_url:

returning a tuple of a ForkedDatabase wrapped in an Arc RwLock and ClientFork wrapped in an Option which can be used in a Backend to fork from.

modifying some parameters of the passed env

mutating some members of self

pub async fn setup_fork_db_config( &mut self, eth_rpc_url: String, env: &mut EnvWithHandlerCfg, fees: &FeeManager, ) -> (ForkedDatabase, ClientForkConfig)

Configures everything related to forking based on the passed eth_rpc_url:

returning a tuple of a ForkedDatabase and ClientForkConfig which can be used to build a ClientFork to fork from.

modifying some parameters of the passed env

mutating some members of self

pub(crate) fn fork_gas_limit( &self, block: &Block<WithOtherFields<Transaction>>, ) -> u128

we only use the gas limit value of the block if it is non-zero and the block gas limit is enabled, since there are networks where this is not used and is always 0x0 which would inevitably result in OutOfGas errors as soon as the evm is about to record gas, See also https://github.com/foundry-rs/foundry/issues/3247

pub(crate) fn gas_limit(&self) -> u128

Returns the gas limit for a non forked anvil instance

Checks the config for the disable_block_gas_limit flag