anvil::eth::backend::db

Trait Db

pub trait Db:
    DatabaseRef<Error = DatabaseError>
    + Database<Error = DatabaseError>
    + DatabaseCommit
    + MaybeFullDatabase
    + MaybeForkedDatabase
    + Debug
    + Send
    + Sync {
    // Required methods
    fn insert_account(&mut self, address: Address, account: AccountInfo);
    fn set_storage_at(
        &mut self,
        address: Address,
        slot: B256,
        val: B256,
    ) -> DatabaseResult<()>;
    fn insert_block_hash(&mut self, number: U256, hash: B256);
    fn dump_state(
        &self,
        at: BlockEnv,
        best_number: u64,
        blocks: Vec<SerializableBlock>,
        transactions: Vec<SerializableTransaction>,
        historical_states: Option<SerializableHistoricalStates>,
    ) -> DatabaseResult<Option<SerializableState>>;
    fn snapshot_state(&mut self) -> U256;
    fn revert_state(
        &mut self,
        state_snapshot: U256,
        action: RevertStateSnapshotAction,
    ) -> bool;
    fn current_state(&self) -> StateDb;

    // Provided methods
    fn set_nonce(&mut self, address: Address, nonce: u64) -> DatabaseResult<()> { ... }
    fn set_balance(
        &mut self,
        address: Address,
        balance: U256,
    ) -> DatabaseResult<()> { ... }
    fn set_code(&mut self, address: Address, code: Bytes) -> DatabaseResult<()> { ... }
    fn load_state(&mut self, state: SerializableState) -> DatabaseResult<bool> { ... }
    fn maybe_state_root(&self) -> Option<B256> { ... }
}

This bundles all required revm traits

Required Methods

fn insert_account(&mut self, address: Address, account: AccountInfo)

Inserts an account

fn set_storage_at( &mut self, address: Address, slot: B256, val: B256, ) -> DatabaseResult<()>

Sets the storage value at the given slot for the address

fn insert_block_hash(&mut self, number: U256, hash: B256)

inserts a blockhash for the given number

fn dump_state( &self, at: BlockEnv, best_number: u64, blocks: Vec<SerializableBlock>, transactions: Vec<SerializableTransaction>, historical_states: Option<SerializableHistoricalStates>, ) -> DatabaseResult<Option<SerializableState>>

Write all chain data to serialized bytes buffer

fn snapshot_state(&mut self) -> U256

Creates a new state snapshot.

fn revert_state( &mut self, state_snapshot: U256, action: RevertStateSnapshotAction, ) -> bool

Reverts a state snapshot.

Returns true if the state snapshot was reverted.

fn current_state(&self) -> StateDb

Returns the current, standalone state of the Db

Provided Methods

fn set_nonce(&mut self, address: Address, nonce: u64) -> DatabaseResult<()>

Sets the nonce of the given address

fn set_balance(&mut self, address: Address, balance: U256) -> DatabaseResult<()>

Sets the balance of the given address

fn set_code(&mut self, address: Address, code: Bytes) -> DatabaseResult<()>

Sets the code of the given address

fn load_state(&mut self, state: SerializableState) -> DatabaseResult<bool>

Deserialize and add all chain data to the backend storage

fn maybe_state_root(&self) -> Option<B256>

Returns the state root if possible to compute

Implementations

impl dyn Db

pub fn as_dbref(&self) -> &dyn DatabaseRef<Error = DatabaseError>