Trait anvil::eth::backend::db::MaybeHashDatabase

pub trait MaybeHashDatabase: DatabaseRef<Error = DatabaseError> {
    // Required methods
    fn clear_into_snapshot(&mut self) -> StateSnapshot;
    fn clear(&mut self);
    fn init_from_snapshot(&mut self, snapshot: StateSnapshot);

    // Provided methods
    fn maybe_as_hash_db(&self) -> Option<(AsHashDB, B256)> { ... }
    fn maybe_account_db(&self, _addr: Address) -> Option<(AsHashDB, B256)> { ... }
}

Helper trait get access to the data in HashDb form

Required Methods

fn clear_into_snapshot(&mut self) -> StateSnapshot

Clear the state and move it into a new StateSnapshot

fn clear(&mut self)

Clears the entire database

fn init_from_snapshot(&mut self, snapshot: StateSnapshot)

Reverses clear_into_snapshot by initializing the db’s state with the snapshot

Provided Methods

fn maybe_as_hash_db(&self) -> Option<(AsHashDB, B256)>

Return the DB as read-only hashdb and the root key

fn maybe_account_db(&self, _addr: Address) -> Option<(AsHashDB, B256)>

Return the storage DB as read-only hashdb and the storage root of the account