anvil::eth::backend

Module executor

Structs

ExecutedTransaction

Represents an executed transaction (transacted on the DB)

ExecutedTransactions

Represents the outcome of mining a new block

TransactionExecutor

An executor for a series of transactions

Enums

TransactionExecutionOutcome

Represents the result of a single transaction execution attempt

Functions

build_logs_bloom 🔒

Inserts all logs into the bloom

new_evm_with_inspector

Creates a database with given database and inspector, optionally enabling odyssey features.

new_evm_with_inspector_ref

Creates a new EVM with the given inspector and wraps the database in a WrapDatabaseRef.