Module anvil::eth::backend::mem

In-memory Backend In-memory blockchain backend.

Modules

cache

fork_db

in_memory_db

The in memory DB

inspector

Anvil specific [revm::Inspector] implementation

state

Support for generating the state root for memdb storage

storage

In-memory blockchain storage

Structs

Backend

Gives access to the [revm::Database]

Enums

BlockRequest

A block request, which includes the Pool Transactions if it’s Pending

Constants

MIN_CREATE_GAS

MIN_TRANSACTION_GAS

Functions

get_pool_transactions_nonce 🔒

Get max nonce from transaction pool by address

prove_storage

Prove a storage key’s existence or nonexistence in the account’s storage trie. storage_key is the hash of the desired storage key, meaning this will only work correctly under a secure trie. storage_key == keccak(key)

transaction_build

Creates a Transaction as it’s expected for the eth RPC api from storage data

Type Aliases

State