anvil::eth::backend::mem

Module state

Support for generating the state root for memdb storage

Functions

apply_block_overrides

Applies the given block overrides to the env and updates overridden block hashes in the db.

apply_state_overrides

Applies the given state overrides to the given CacheDB

build_root

state_root

Builds state root from the given accounts

storage_root

Builds storage root from the given storage

trie_account_rlp

Returns the RLP for this account.

trie_accounts

Builds iterator over stored key-value pairs ready for account trie root calculation.

trie_storage

Builds iterator over stored key-value pairs ready for storage trie root calculation.