anvil::eth::backend::mem
Struct Backend

pub struct Backend {

    db: Arc<RwLock<Box<dyn Db>>>,
    blockchain: Blockchain,
    states: Arc<RwLock<InMemoryBlockStates>>,
    env: Arc<RwLock<EnvWithHandlerCfg>>,
    fork: Arc<RwLock<Option<ClientFork>>>,
    time: TimeManager,
    cheats: CheatsManager,
    fees: FeeManager,
    genesis: GenesisConfig,
    new_block_listeners: Arc<Mutex<Vec<UnboundedSender<NewBlockNotification>>>>,
    active_state_snapshots: Arc<Mutex<HashMap<U256, (u64, B256)>>>,
    enable_steps_tracing: bool,
    print_logs: bool,
    alphanet: bool,
    prune_state_history_config: PruneStateHistoryConfig,
    transaction_block_keeper: Option<usize>,
    node_config: Arc<RwLock<NodeConfig>>,
    slots_in_an_epoch: u64,
    precompile_factory: Option<Arc<dyn PrecompileFactory>>,
    mining: Arc<Mutex<()>>,
    capabilities: Arc<RwLock<WalletCapabilities>>,
    executor_wallet: Arc<RwLock<Option<EthereumWallet>>>,

}

Gives access to the [revm::Database]

Fields
db: Arc<RwLock<Box<dyn Db>>>

Access to [revm::Database] abstraction.

This will be used in combination with [revm::Evm] and is responsible for feeding data to the evm during its execution.

At time of writing, there are two different types of Db:

MemDb: everything is stored in memory

ForkDb: forks off a remote client, missing data is retrieved via RPC-calls

In order to commit changes to the [revm::Database], the [revm::Evm] requires mutable access, which requires a write-lock from this db. In forking mode, the time during which the write-lock is active depends on whether the ForkDb can provide all requested data from memory or whether it has to retrieve it via RPC calls first. This means that it potentially blocks for some time, even taking into account the rate limits of RPC endpoints. Therefore the Db is guarded by a tokio::sync::RwLock here so calls that need to read from it, while it’s currently written to, don’t block. E.g. a new block is currently mined and a new Self::set_storage_at() request is being executed.

blockchain: Blockchain

stores all block related data in memory.

states: Arc<RwLock<InMemoryBlockStates>>

Historic states of previous blocks.

env: Arc<RwLock<EnvWithHandlerCfg>>

Env data of the chain

fork: Arc<RwLock<Option<ClientFork>>>

This is set if this is currently forked off another client.

time: TimeManager

Provides time related info, like timestamp.

cheats: CheatsManager

Contains state of custom overrides.

fees: FeeManager

Contains fee data.

genesis: GenesisConfig

Initialised genesis.

new_block_listeners: Arc<Mutex<Vec<UnboundedSender<NewBlockNotification>>>>

Listeners for new blocks that get notified when a new block was imported.

active_state_snapshots: Arc<Mutex<HashMap<U256, (u64, B256)>>>

Keeps track of active state snapshots at a specific block.

enable_steps_tracing: boolprint_logs: boolalphanet: boolprune_state_history_config: PruneStateHistoryConfig

How to keep history state

transaction_block_keeper: Option<usize>

max number of blocks with transactions in memory

node_config: Arc<RwLock<NodeConfig>>slots_in_an_epoch: u64

Slots in an epoch

precompile_factory: Option<Arc<dyn PrecompileFactory>>

Precompiles to inject to the EVM.

mining: Arc<Mutex<()>>

Prevent race conditions during mining

capabilities: Arc<RwLock<WalletCapabilities>>executor_wallet: Arc<RwLock<Option<EthereumWallet>>>
Implementations

impl Backend

pub async fn with_genesis( db: Arc<AsyncRwLock<Box<dyn Db>>>, env: Arc<RwLock<EnvWithHandlerCfg>>, genesis: GenesisConfig, fees: FeeManager, fork: Arc<RwLock<Option<ClientFork>>>, enable_steps_tracing: bool, print_logs: bool, alphanet: bool, prune_state_history_config: PruneStateHistoryConfig, max_persisted_states: Option<usize>, transaction_block_keeper: Option<usize>, automine_block_time: Option<Duration>, node_config: Arc<AsyncRwLock<NodeConfig>>, ) -> Self

Initialises the balance of the given accounts

pub async fn set_create2_deployer(&self, address: Address) -> DatabaseResult<()>

Writes the CREATE2 deployer code directly to the database at the address provided.

pub(crate) fn get_capabilities(&self) -> WalletCapabilities

Get the capabilities of the wallet.

Currently the only capability is DelegationCapability.

pub(crate) fn update_interval_mine_block_time(&self, block_time: Duration)

Updates memory limits that should be more strict when auto-mine is enabled

pub(crate) fn executor_wallet(&self) -> Option<EthereumWallet>

pub(crate) fn add_capability(&self, address: Address)

Adds an address to the [DelegationCapability] of the wallet.

pub(crate) fn set_executor( &self, executor_pk: String, ) -> Result<Address, BlockchainError>

async fn apply_genesis(&self) -> Result<(), DatabaseError>

Applies the configured genesis settings

This will fund, create the genesis accounts

pub fn impersonate(&self, addr: Address) -> bool

Sets the account to impersonate

Returns true if the account is already impersonated

pub fn stop_impersonating(&self, addr: Address)

Removes the account that from the impersonated set

If the impersonated addr is a contract then we also reset the code here

pub fn auto_impersonate_account(&self, enabled: bool)

If set to true will make every account impersonated

pub fn get_fork(&self) -> Option<ClientFork>

Returns the configured fork, if any

pub fn get_db(&self) -> &Arc<AsyncRwLock<Box<dyn Db>>>

Returns the database

pub async fn get_account(&self, address: Address) -> DatabaseResult<AccountInfo>

Returns the AccountInfo from the database

pub fn is_fork(&self) -> bool

Whether we’re forked off some remote client

pub fn precompiles(&self) -> Vec<Address>

pub async fn reset_fork(&self, forking: Forking) -> Result<(), BlockchainError>

Resets the fork to a fresh state

pub fn time(&self) -> &TimeManager

Returns the TimeManager responsible for timestamps

pub fn cheats(&self) -> &CheatsManager

Returns the CheatsManager responsible for executing cheatcodes

pub fn fees(&self) -> &FeeManager

Returns the FeeManager that manages fee/pricings

pub fn env(&self) -> &Arc<RwLock<EnvWithHandlerCfg>>

The env data of the blockchain

pub fn best_hash(&self) -> B256

Returns the current best hash of the chain

pub fn best_number(&self) -> u64

Returns the current best number of the chain

pub fn set_block_number(&self, number: U256)

Sets the block number

pub fn coinbase(&self) -> Address

Returns the client coinbase address.

pub fn chain_id(&self) -> U256

Returns the client coinbase address.

pub fn set_chain_id(&self, chain_id: u64)

pub async fn current_balance(&self, address: Address) -> DatabaseResult<U256>

Returns balance of the given account.

pub async fn current_nonce(&self, address: Address) -> DatabaseResult<u64>

Returns balance of the given account.

pub fn set_coinbase(&self, address: Address)

Sets the coinbase address

pub async fn set_nonce( &self, address: Address, nonce: U256, ) -> DatabaseResult<()>

Sets the nonce of the given address

pub async fn set_balance( &self, address: Address, balance: U256, ) -> DatabaseResult<()>

Sets the balance of the given address

pub async fn set_code( &self, address: Address, code: Bytes, ) -> DatabaseResult<()>

Sets the code of the given address

pub async fn set_storage_at( &self, address: Address, slot: U256, val: B256, ) -> DatabaseResult<()>

Sets the value for the given slot of the given address

pub fn spec_id(&self) -> SpecId

Returns the configured specid

pub fn is_eip1559(&self) -> bool

Returns true for post London

pub fn is_eip3675(&self) -> bool

Returns true for post Merge

pub fn is_eip2930(&self) -> bool

Returns true for post Berlin

pub fn is_eip4844(&self) -> bool

Returns true for post Cancun

pub fn is_eip7702(&self) -> bool

Returns true for post Prague

pub fn is_optimism(&self) -> bool

Returns true if op-stack deposits are active

pub fn ensure_eip1559_active(&self) -> Result<(), BlockchainError>

Returns an error if EIP1559 is not active (pre Berlin)

pub fn ensure_eip2930_active(&self) -> Result<(), BlockchainError>

Returns an error if EIP1559 is not active (pre muirGlacier)

pub fn ensure_eip4844_active(&self) -> Result<(), BlockchainError>

pub fn ensure_eip7702_active(&self) -> Result<(), BlockchainError>

pub fn ensure_op_deposits_active(&self) -> Result<(), BlockchainError>

Returns an error if op-stack deposits are not active

pub fn gas_limit(&self) -> u128

Returns the block gas limit

pub fn set_gas_limit(&self, gas_limit: u128)

Sets the block gas limit

pub fn base_fee(&self) -> u64

Returns the current base fee

pub fn is_min_priority_fee_enforced(&self) -> bool

Returns whether the minimum suggested priority fee is enforced

pub fn excess_blob_gas_and_price(&self) -> Option<BlobExcessGasAndPrice>

pub fn set_base_fee(&self, basefee: u64)

Sets the current basefee

pub fn set_gas_price(&self, price: u128)

Sets the gas price

pub fn elasticity(&self) -> f64

pub fn total_difficulty(&self) -> U256

Returns the total difficulty of the chain until this block

Note: this will always be 0 in memory mode In forking mode this will always be the total difficulty of the forked block

pub async fn create_state_snapshot(&self) -> U256

Creates a new evm_snapshot at the current height.

Returns the id of the snapshot created.

pub async fn revert_state_snapshot( &self, id: U256, ) -> Result<bool, BlockchainError>

Reverts the state to the state snapshot identified by the given id.

pub fn list_state_snapshots(&self) -> BTreeMap<U256, (u64, B256)>

pub async fn serialized_state( &self, preserve_historical_states: bool, ) -> Result<SerializableState, BlockchainError>

Get the current state.

pub async fn dump_state( &self, preserve_historical_states: bool, ) -> Result<Bytes, BlockchainError>

Write all chain data to serialized bytes buffer

pub async fn load_state( &self, state: SerializableState, ) -> Result<bool, BlockchainError>

Apply SerializableState data to the backend storage.

pub async fn load_state_bytes( &self, buf: Bytes, ) -> Result<bool, BlockchainError>

Deserialize and add all chain data to the backend storage

fn next_env(&self) -> EnvWithHandlerCfg

Returns the environment for the next block

fn new_evm_with_inspector_ref<'i, 'db>( &self, db: &'db dyn DatabaseRef<Error = DatabaseError>, env: EnvWithHandlerCfg, inspector: &'i mut dyn Inspector<WrapDatabaseRef<&'db dyn DatabaseRef<Error = DatabaseError>>>, ) -> Evm<'_, &'i mut dyn Inspector<WrapDatabaseRef<&'db dyn DatabaseRef<Error = DatabaseError>>>, WrapDatabaseRef<&'db dyn DatabaseRef<Error = DatabaseError>>>

Creates an EVM instance with optionally injected precompiles.

pub async fn inspect_tx( &self, tx: Arc<PoolTransaction>, ) -> Result<(InstructionResult, Option<Output>, u64, State, Vec<Log>), BlockchainError>

executes the transactions without writing to the underlying database

pub async fn pending_block( &self, pool_transactions: Vec<Arc<PoolTransaction>>, ) -> BlockInfo

Creates the pending block

This will execute all transaction in the order they come but will not mine the block

pub async fn with_pending_block<F, T>( &self, pool_transactions: Vec<Arc<PoolTransaction>>, f: F, ) -> T
where F: FnOnce(Box<dyn MaybeFullDatabase + '_>, BlockInfo) -> T,

Creates the pending block

This will execute all transaction in the order they come but will not mine the block

pub async fn mine_block( &self, pool_transactions: Vec<Arc<PoolTransaction>>, ) -> MinedBlockOutcome

Mines a new block and stores it.

this will execute all transaction in the order they come in and return all the markers they provide.

async fn do_mine_block( &self, pool_transactions: Vec<Arc<PoolTransaction>>, ) -> MinedBlockOutcome

pub async fn call( &self, request: WithOtherFields<TransactionRequest>, fee_details: FeeDetails, block_request: Option<BlockRequest>, overrides: Option<StateOverride>, ) -> Result<(InstructionResult, Option<Output>, u128, State), BlockchainError>

Executes the [TransactionRequest] without writing to the DB

Errors

Returns an error if the block_number is greater than the current height

fn build_call_env( &self, request: WithOtherFields<TransactionRequest>, fee_details: FeeDetails, block_env: BlockEnv, ) -> EnvWithHandlerCfg

EVM settings

This modifies certain EVM settings to mirror geth’s SkipAccountChecks when transacting requests, see also: https://github.com/ethereum/go-ethereum/blob/380688c636a654becc8f114438c2a5d93d2db032/core/state_transition.go#L145-L148:

disable_eip3607 is set to true

disable_base_fee is set to true

nonce is set to None

fn build_inspector(&self) -> Inspector

Builds Inspector with the configured options

pub fn call_with_state( &self, state: &dyn DatabaseRef<Error = DatabaseError>, request: WithOtherFields<TransactionRequest>, fee_details: FeeDetails, block_env: BlockEnv, ) -> Result<(InstructionResult, Option<Output>, u128, State), BlockchainError>

pub async fn call_with_tracing( &self, request: WithOtherFields<TransactionRequest>, fee_details: FeeDetails, block_request: Option<BlockRequest>, opts: GethDebugTracingCallOptions, ) -> Result<GethTrace, BlockchainError>

pub fn build_access_list_with_state( &self, state: &dyn DatabaseRef<Error = DatabaseError>, request: WithOtherFields<TransactionRequest>, fee_details: FeeDetails, block_env: BlockEnv, ) -> Result<(InstructionResult, Option<Output>, u64, AccessList), BlockchainError>

fn get_receipts( &self, tx_hashes: impl IntoIterator<Item = TxHash>, ) -> Vec<TypedReceipt>

returns all receipts for the given transactions

async fn logs_for_block( &self, filter: Filter, hash: B256, ) -> Result<Vec<Log>, BlockchainError>

Returns the logs of the block that match the filter

fn mined_logs_for_block(&self, filter: Filter, block: Block) -> Vec<Log>

Returns all Logs mined by the node that were emitted in the block and match the Filter

async fn logs_for_range( &self, filter: &Filter, from: u64, to: u64, ) -> Result<Vec<Log>, BlockchainError>

Returns the logs that match the filter in the given range of blocks

pub async fn logs(&self, filter: Filter) -> Result<Vec<Log>, BlockchainError>

Returns the logs according to the filter

pub async fn block_by_hash( &self, hash: B256, ) -> Result<Option<AnyNetworkBlock>, BlockchainError>

pub async fn block_by_hash_full( &self, hash: B256, ) -> Result<Option<AnyNetworkBlock>, BlockchainError>

fn mined_block_by_hash(&self, hash: B256) -> Option<AnyNetworkBlock>

pub(crate) async fn mined_transactions_by_block_number( &self, number: BlockNumber, ) -> Option<Vec<WithOtherFields<Transaction>>>

pub(crate) fn mined_transactions_in_block( &self, block: &Block, ) -> Option<Vec<WithOtherFields<Transaction>>>

Returns all transactions given a block

pub async fn block_by_number( &self, number: BlockNumber, ) -> Result<Option<AnyNetworkBlock>, BlockchainError>

pub async fn block_by_number_full( &self, number: BlockNumber, ) -> Result<Option<AnyNetworkBlock>, BlockchainError>

pub fn get_block(&self, id: impl Into<BlockId>) -> Option<Block>

pub fn get_block_by_hash(&self, hash: B256) -> Option<Block>

pub fn mined_block_by_number( &self, number: BlockNumber, ) -> Option<AnyNetworkBlock>

pub fn get_full_block(&self, id: impl Into<BlockId>) -> Option<AnyNetworkBlock>

pub fn convert_block(&self, block: Block) -> AnyNetworkBlock

Takes a block as it’s stored internally and returns the eth api conform block format.

pub async fn ensure_block_number<T: Into<BlockId>>( &self, block_id: Option<T>, ) -> Result<u64, BlockchainError>

Converts the BlockNumber into a numeric value

Errors

returns an error if the requested number is larger than the current height

pub fn convert_block_number(&self, block: Option<BlockNumber>) -> u64

pub async fn with_database_at<F, T>( &self, block_request: Option<BlockRequest>, f: F, ) -> Result<T, BlockchainError>
where F: FnOnce(Box<dyn MaybeFullDatabase + '_>, BlockEnv) -> T,

Helper function to execute a closure with the database at a specific block

pub async fn storage_at( &self, address: Address, index: U256, block_request: Option<BlockRequest>, ) -> Result<B256, BlockchainError>

pub async fn get_code( &self, address: Address, block_request: Option<BlockRequest>, ) -> Result<Bytes, BlockchainError>

Returns the code of the address

If the code is not present and fork mode is enabled then this will try to fetch it from the forked client

pub fn get_code_with_state( &self, state: &dyn DatabaseRef<Error = DatabaseError>, address: Address, ) -> Result<Bytes, BlockchainError>

pub async fn get_balance( &self, address: Address, block_request: Option<BlockRequest>, ) -> Result<U256, BlockchainError>

Returns the balance of the address

If the requested number predates the fork then this will fetch it from the endpoint

pub async fn get_account_at_block( &self, address: Address, block_request: Option<BlockRequest>, ) -> Result<Account, BlockchainError>

pub fn get_balance_with_state<D>( &self, state: D, address: Address, ) -> Result<U256, BlockchainError>
where D: DatabaseRef<Error = DatabaseError>,

pub async fn get_nonce( &self, address: Address, block_request: BlockRequest, ) -> Result<u64, BlockchainError>

Returns the nonce of the address

If the requested number predates the fork then this will fetch it from the endpoint

pub async fn trace_transaction( &self, hash: B256, ) -> Result<Vec<LocalizedTransactionTrace>, BlockchainError>

Returns the traces for the given transaction

pub(crate) fn mined_parity_trace_transaction( &self, hash: B256, ) -> Option<Vec<LocalizedTransactionTrace>>

Returns the traces for the given transaction

pub(crate) fn mined_transaction(&self, hash: B256) -> Option<MinedTransaction>

Returns the traces for the given transaction

pub(crate) fn mined_parity_trace_block( &self, block: u64, ) -> Option<Vec<LocalizedTransactionTrace>>

Returns the traces for the given block

pub async fn debug_trace_transaction( &self, hash: B256, opts: GethDebugTracingOptions, ) -> Result<GethTrace, BlockchainError>

Returns the traces for the given transaction

fn mined_geth_trace_transaction( &self, hash: B256, opts: GethDebugTracingOptions, ) -> Option<Result<GethTrace, BlockchainError>>

pub async fn trace_block( &self, block: BlockNumber, ) -> Result<Vec<LocalizedTransactionTrace>, BlockchainError>

Returns the traces for the given block

pub async fn transaction_receipt( &self, hash: B256, ) -> Result<Option<ReceiptResponse>, BlockchainError>

pub async fn trace_filter( &self, filter: TraceFilter, ) -> Result<Vec<LocalizedTransactionTrace>, BlockchainError>

pub fn mined_receipts(&self, hash: B256) -> Option<Vec<TypedReceipt>>

Returns all receipts of the block

pub fn mined_block_receipts( &self, id: impl Into<BlockId>, ) -> Option<Vec<ReceiptResponse>>

Returns all transaction receipts of the block

pub(crate) fn mined_transaction_receipt( &self, hash: B256, ) -> Option<MinedTransactionReceipt>

Returns the transaction receipt for the given hash

pub async fn block_receipts( &self, number: BlockId, ) -> Result<Option<Vec<ReceiptResponse>>, BlockchainError>

Returns the blocks receipts for the given number

pub async fn transaction_by_block_number_and_index( &self, number: BlockNumber, index: Index, ) -> Result<Option<WithOtherFields<Transaction>>, BlockchainError>

pub async fn transaction_by_block_hash_and_index( &self, hash: B256, index: Index, ) -> Result<Option<WithOtherFields<Transaction>>, BlockchainError>

pub fn mined_transaction_by_block_hash_and_index( &self, block_hash: B256, index: Index, ) -> Option<WithOtherFields<Transaction>>

pub async fn transaction_by_hash( &self, hash: B256, ) -> Result<Option<WithOtherFields<Transaction>>, BlockchainError>

pub fn mined_transaction_by_hash( &self, hash: B256, ) -> Option<WithOtherFields<Transaction>>

pub async fn prove_account_at( &self, address: Address, keys: Vec<B256>, block_request: Option<BlockRequest>, ) -> Result<AccountProof, BlockchainError>

Prove an account’s existence or nonexistence in the state trie.

Returns a merkle proof of the account’s trie node, account_key == keccak(address)

pub fn new_block_notifications(&self) -> NewBlockNotifications

Returns a new block event stream

fn notify_on_new_block(&self, header: Header, hash: B256)

Notifies all new_block_listeners about the new block

pub async fn reorg( &self, depth: u64, tx_pairs: HashMap<u64, Vec<Arc<PoolTransaction>>>, common_block: Block, ) -> Result<(), BlockchainError>

Reorg the chain to a common height and execute blocks to build new chain.

The state of the chain is rewound using rewind to the common block, including the db, storage, and env.

Finally, do_mine_block is called to create the new chain.