Module anvil::eth::otterscan::types

Structs

Issuance

Issuance information for a block. Expected by Otterscan in ots_getBlockDetails calls

OtsBlock

Patched Block struct, to include the additional transactionCount field expected by Otterscan

OtsBlockDetails

Block structure with additional details regarding fees and issuance

OtsBlockTransactions

Holds both transactions and receipts for a block

OtsContractCreator

Information about the creator address and transaction for a contract

OtsInternalOperation

Otterscan format for listing relevant internal operations.

OtsSearchTransactions

Paginated search results of an account’s history

OtsTrace

Otterscan’s representation of a trace

OtsTransactionReceipt

Patched Receipt struct, to include the additional timestamp field expected by Otterscan

Enums

OtsInternalOperationType

Types of internal operations recognized by Otterscan.

OtsTraceType

The type of call being described by an Otterscan trace. Only CALL, STATICCALL and DELEGATECALL are represented