anvil

Module filter

support for polling filters Support for polling based filters

Structs

Filters

Contains all registered filters

LogsFilter

Listens for new blocks and matching logs emitted in that block

Enums

EthFilter

Represents a poll based filter

Constants

ACTIVE_FILTER_TIMEOUT_SECS

timeout after which to remove an active filter if it wasn’t polled since then

Functions

new_id 🔒

returns a new random hex id

Type Aliases

FilterMap 🔒

Type alias for filters identified by their id and their expiration timestamp