anvil

Module server

axum RPC server implementations Contains the code to launch an Ethereum RPC server.

Modules

error

handler 🔒

Contains RPC handlers

Functions

router

Configures an [axum::Router] that handles EthApi related JSON-RPC calls via HTTP and WS.

serve

Configures a server that handles EthApi related JSON-RPC calls via HTTP and WS.

serve_on

Configures a server that handles EthApi related JSON-RPC calls via HTTP and WS.

spawn_ipc

Launches an ipc server at the given path in a new task

try_spawn_ipc

Launches an ipc server at the given path in a new task.