Struct anvil::service::BlockProducer

struct BlockProducer {
    idle_backend: Option<Arc<Backend>>,
    block_mining: Option<JoinHandle<(MinedBlockOutcome, Arc<Backend>)>>,
    queued: VecDeque<Vec<Arc<PoolTransaction>>>,
}

A type that exclusively mines one block at a time

Fields
idle_backend: Option<Arc<Backend>>

Holds the backend if no block is being mined

block_mining: Option<JoinHandle<(MinedBlockOutcome, Arc<Backend>)>>

Single active future that mines a new block

queued: VecDeque<Vec<Arc<PoolTransaction>>>

backlog of sets of transactions ready to be mined

Implementations

impl BlockProducer

fn new(backend: Arc<Backend>) -> Self