anvil

Trait PrecompileFactory

pub trait PrecompileFactory:
    Send
    + Sync
    + Unpin
    + Debug {
    // Required method
    fn precompiles(&self) -> Vec<(Precompile, u64)>;
}

Object-safe trait that enables injecting extra precompiles when using anvil as a library.

Required Methods

fn precompiles(&self) -> Vec<(Precompile, u64)>

Returns a set of precompiles to extend the EVM with.