anvil

Module config

Structs

AccountGenerator

Can create dev accounts

NodeConfig

Configurations of the EVM node

PruneStateHistoryConfig

Enums

ForkChoice

Fork delimiter used to specify which block or transaction to fork from.

Constants

BANNER 🔒

CHAIN_ID

Default chain id of the node

DEFAULT_GAS_LIMIT

The default gas limit for all transactions

DEFAULT_IPC_ENDPOINT

The default IPC endpoint

DEFAULT_MNEMONIC

Default mnemonic for dev accounts

NODE_PORT

Default port the rpc will open

VERSION_MESSAGE

The short version message information for the Foundry CLI.

Functions

anvil_dir

Returns the path to anvil dir ~/.foundry/anvil

anvil_tmp_dir

Returns the root path to anvil’s temporary storage ~/.foundry/anvil/

derive_block_and_transactions 🔒

If the fork choice is a block number, simply return it with an empty list of transactions. If the fork choice is a transaction hash, determine the block that the transaction was mined in, and return the block number before the fork block along with all transactions in the fork block that are before (and including) the fork transaction.

find_latest_fork_block 🔒

Finds the latest appropriate block to fork