Module anvil::eth::api

Structs

EthApi

The entry point for executing eth api RPC call - The Eth RPC interface.

Enums

GasEstimationCallResult 🔒

Keeps result of a call to revm EVM used for gas estimation

Constants

CLIENT_VERSION

The client version: anvil/v{major}.{minor}.{patch}

Functions

convert_transact_out 🔒

determine_base_gas_by_kind 🔒

Determines the minimum gas needed for a transaction depending on the transaction kind.

ensure_return_ok 🔒

Returns an error if the exit code is not ok

required_marker 🔒