Trait anvil::eth::backend::db::MaybeFullDatabase

pub trait MaybeFullDatabase: DatabaseRef<Error = DatabaseError> {
    // Required methods
    fn clear_into_snapshot(&mut self) -> StateSnapshot;
    fn read_as_snapshot(&self) -> StateSnapshot;
    fn clear(&mut self);
    fn init_from_snapshot(&mut self, snapshot: StateSnapshot);

    // Provided method
    fn maybe_as_full_db(&self) -> Option<&HashMap<Address, DbAccount>> { ... }
}

Helper trait get access to the full state data of the database

Required Methods

fn clear_into_snapshot(&mut self) -> StateSnapshot

Clear the state and move it into a new StateSnapshot

fn read_as_snapshot(&self) -> StateSnapshot

Read the state snapshot

This clones all the states and returns a new StateSnapshot

fn clear(&mut self)

Clears the entire database

fn init_from_snapshot(&mut self, snapshot: StateSnapshot)

Reverses clear_into_snapshot by initializing the db’s state with the snapshot

Provided Methods

fn maybe_as_full_db(&self) -> Option<&HashMap<Address, DbAccount>>