anvil::eth::backend
Module mem

In-memory Backend In-memory blockchain backend.

Modules

cache

fork_db

in_memory_db

The in memory DB

inspector

Anvil specific [revm::Inspector] implementation

state

Support for generating the state root for memdb storage

storage

In-memory blockchain storage

Structs

Backend

Gives access to the [revm::Database]

Enums

BlockRequest

A block request, which includes the Pool Transactions if it’s Pending

Constants

EXECUTOR

EXECUTOR_PK

EXP_ERC20_CONTRACT

EXP_ERC20_RUNTIME_CODE

MIN_CREATE_GAS

MIN_TRANSACTION_GAS

P256_DELEGATION_CONTRACT

P256_DELEGATION_RUNTIME_CODE

Functions

get_pool_transactions_nonce 🔒

Get max nonce from transaction pool by address

is_arbitrum

prove_storage

Prove a storage key’s existence or nonexistence in the account’s storage trie.

transaction_build

Creates a Transaction as it’s expected for the eth RPC api from storage data

Type Aliases

State